api_service/requirements.txt
Flask
Flask-Cors
Flask-Session
Flask-Blueprint
Flask-Bcrypt
python-dotenv
werkzeug
google-cloud-firestore
requests
firebase_admin
oauthlib
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests-oauthlib==2.0.0
jsonschema 